requirements.txt
gradio>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
qrcode>=8.2.0
torch>=2.0.0
transformers>=4.52.0
accelerate>=0.20.0
anthropic>=0.52.2
safetensors>=0.4.3
einops>=0.6.1
tokenizers>=0.13.3
requests>=2.31.0
tqdm>=4.65.0
huggingface-hub>=0.17.0
opencv-python>=4.8.0
scikit-image>=0.21.0
